Restructure sender and nonces management
