+++ utils/node-loader/src/batch_pool/batch/mod.rs
@@ -1,0 +1,53 @@
+use super::Seed;
+
+pub use upload_code::{UploadCodeArgs, UploadCodeBatchOutput};
+pub use upload_program::{UploadProgramArgs, UploadProgramArgsInner, UploadProgramBatchOutput};
+pub use create_program::{CreateProgramArgs, CreateProgramArgsInner, CreateProgramBatchOutput};
+pub use send_message::{SendMessageArgs, SendMessageArgsInner, SendMessageBatchOutput};
+
+mod create_program;
+mod send_message;
+mod upload_code;
+mod upload_program;
+
+
+pub enum Batch {
+    UploadProgram(Vec<UploadProgramArgs>),
+    UploadCode(Vec<UploadCodeArgs>),
+    SendMessage(Vec<SendMessageArgs>),
+    CreateProgram(Vec<CreateProgramArgs>),
+}
+
+pub struct BatchWithSeed {
+    pub seed: Seed,
+    pub batch: Batch,
+}
+
+impl BatchWithSeed {
+    pub fn batch_str(&self) -> &'static str {
+        match &self.batch {
+            Batch::UploadProgram(_) => "upload_program",
+            Batch::UploadCode(_) => "upload_code",
+            Batch::SendMessage(_) => "send_message",
+            Batch::CreateProgram(_) => "create_program",
+        }
+    }
+}
+
+impl From<BatchWithSeed> for Batch {
+    fn from(other: BatchWithSeed) -> Self {
+        other.batch
+    }
+}
+
+impl From<(Seed, Batch)> for BatchWithSeed {
+    fn from((seed, batch): (Seed, Batch)) -> Self {
+        Self { seed, batch }
+    }
+}
+
+impl From<BatchWithSeed> for (Seed, Batch) {
+    fn from(BatchWithSeed { seed, batch }: BatchWithSeed) -> Self {
+        (seed, batch)
+    }
+}
